+++ cards/_components/Dashboard/FancyModal.vue
