+++ cards/_components/Dashboard/Table.vue
